+++ TractorMod/Framework/Config/ScytheConfig.cs
@@ -24,8 +24,11 @@ namespace Pathoschild.Stardew.TractorMod.Framework.Config
         /// <summary>Whether to collect machine output.</summary>
         public bool HarvestMachines { get; set; } = false;
 
-        /// <summary>Whether to cut tall grass.</summary>
-        public bool HarvestGrass { get; set; } = true;
+        /// <summary>Whether to cut non-blue tall grass.</summary>
+        public bool HarvestNonBlueGrass { get; set; } = true;
+
+        /// <summary>Whether to cut blue grass.</summary>
+        public bool HarvestBlueGrass { get; set; } = true;
 
         /// <summary>Whether to clear dead crops.</summary>
         public bool ClearDeadCrops { get; set; } = true;