+++ TractorMod/docs/release-notes.md
@@ -5,6 +5,7 @@
 * Added support for...
   * harvesting tree moss (with a scythe);
   * planting blue grass starter.
+* Split 'harvest grass' option into non-blue and blue grass.
 * Simplified tractor names to improve display in mods like NPC Map Locations.
 * Fixed tool options:
   * Fixed axe 'chop free trees (saplings)' option affecting full-grown trees.